Initialize Glib threads in LogStream::Priv.
This means that people do not need to call Initializer::do_init()
in their code. Also checking if the threading system has been
initialized is a cheap operation (see comment in the code), so
it's ok. do_init() is still there although it's modified to check
if the system has been previously initialized.

The whole Initializer is likely to be removed soon actually.