Changed GTKMM_UTILS_API_VERSION to hardcoded 1.0.
This small change makes a lot of things clearer. The API version
should not really be changing after every release. The API itself
does change of course, but the API version number (which is not
reached and stable, but a representation of an aim) should stay
the same. That is, when we declare API stable, and then after that
perhaps break it for some reason, it is then when this number
will change.

By the way, this also means that the headers will be installed
in $prefix/include/gtkmm-utils-1.0/gtkmm-utils/ in each release.
Previously this would be a different root directory every time,
which is not practical, creates problems for packaging etc.