TileView: keep Tiles in TileData structs, added basic pagination-related methods.
Removed the iterator accessors, a shielding for_each_tile(SlotForEachTile)
should be used for iterating through all child Tiles.

The actual pagination code is yet to be done. Note that this means the people
will be unable to pass around Gtk::managed pointers, as during the page viewing
tiles will be thrown in and out of a TileView. So this should be documented l8r.