Added uprintf(), page information in TilePageNavigator.
uprintf() is a printf for Glib::ustrings. We need it now for i18n-ready
strings, because it is still early to require the compose API in just
released glibmm.

Added a separate label for page information text in TPN, ie x - y of z.
TileView updates it in update_navigator_page_info_label(), which should
be called whenever a new Tile is added, or a page change is requested.