Tiles: added option to make the labels aligned to the image.
The behaviour is available via a ctor param.

Commented the ellipsize-ing lines as those now have no effect.

We achieve having the labels left aligned as packed in HBoxes.
The really nasty thing is that, if a label is set to ellipsize,
and without some maximum width explicitly set, *and* packed
indirectly via an HBox, the text will always appear just as "...".

I still don't quite understand why, but in any case I now see the
option to have everything left-aligned absolutely necessary, as having
many widgets (text labels) nicely aligned in the center (by the very
nature of central alignment) is impossible; they all get evenly
distributed in best case. OTOH something like "start kind of from
the center, but continue packing not-expanding is not possible with GTK+.

Also set to have some more spacing between the image and the description.